check for existence of a collection before deleting/updating it

+++ src/routes/collection/[collectionId]/+server.ts
@@ -1,4 +1,4 @@
-import { json } from "@sveltejs/kit";
+import { json, error } from "@sveltejs/kit";
 import prismaClient from "$lib/database";
 
 export const GET = async ({ params }) => {
@@ -14,8 +14,16 @@ export const GET = async ({ params }) => {
 
 export const PATCH = async ({ request, params }) => {
   const collectionId = parseInt(params.collectionId);
+  const collection = await prismaClient.collection.findUnique({
+    where: {
+      id: collectionId,
+    },
+  });
+  if (!collection) {
+    return error(404, "Collection not found");
+  }
   const { open, quests, name, description, owner, ownerId } = await request.json();
-  const collection = await prismaClient.collection.update({
+  const updatedCollection = await prismaClient.collection.update({
     data: {
       open,
       quests: {
@@ -31,11 +39,19 @@ export const PATCH = async ({ request, params }) => {
     },
     include: { quests: { select: { id: true } } },
   });
-  return json(collection, { status: 201 });
+  return json(updatedCollection, { status: 201 });
 };
 
 export const DELETE = async ({ params }) => {
   const collectionId = parseInt(params.collectionId);
+  const collection = await prismaClient.collection.findUnique({
+    where: {
+      id: collectionId,
+    },
+  });
+  if (!collection) {
+    return error(404, "Collection not found");
+  }
   await prismaClient.collection.delete({
     where: {
       id: collectionId,
